requirements.txt
nonebot2[fastapi]
nonebot-adapter-console
nonebot-adapter-onebot

aiohttp
redis
requests
